fix sanitize

+++ include/parsing.h
@@ -17,10 +17,10 @@
 
 typedef struct s_rgb_color
 {
-	int		r;
-	int		g;
-	int		b;
-	int		a;
+	uint8_t	r;
+	uint8_t	g;
+	uint8_t	b;
+	uint8_t	a;
 }	t_rgb_color;
 
 // enums
